packaging: arch: add Arch Linux packaging support
Add PKGBUILD into kci-dev, this builds kci-dev and publishes the
*.pkg.tar.zst into GitHub release.

We extract the version from pyproject.toml for CI to automate release
versions. This possibly can be a cleaner way of doing it, but it gets us
half way there.

We can add steps to auto publish to AUR, or simply let users install it
from the GitHub release.

Signed-off-by: Ben Copeland <ben.copeland@linaro.org>